week2/Airflow/requirements.txt
pandas
pyarrow
numpy
postgres
apache-airflow[google]
sqlalchemy
astro-run-dag # This package is needed for the astro run command. It will be removed before a deploy